Issue 145: Move controllers in the API port Controllers are an adapter of the API port
