+++ back/composer.json
@@ -50,9 +50,9 @@
   "autoload-dev": {
     "psr-4": {
       "App\\Tests\\Acceptance\\": "tests/acceptance/",
+      "App\\Tests\\EndToEnd\\": "tests/end-to-end/",
       "App\\Tests\\Fixtures\\": "tests/fixtures/",
-      "App\\Tests\\Integration\\": "tests/integration/",
-      "App\\Tests\\EndToEnd\\": "tests/end-to-end/"
+      "App\\Tests\\Integration\\": "tests/integration/"
     }
   },
   "replace": {
