+++ back/config/routes/annotations.yaml
@@ -1,3 +1,3 @@
 controllers:
-    resource: ../../src/Infrastructure/Controller/
+    resource: ../../src/Infrastructure/API/Controller/
     type: annotation
