+++ back/config/services.yaml
@@ -20,8 +20,8 @@ services:
 
     # controllers are imported separately to make sure services can be injected
     # as action arguments even if you don't extend any base controller class
-    App\Infrastructure\Controller\:
-        resource: '../src/Infrastructure/Controller'
+    App\Infrastructure\API\Controller\:
+        resource: '../src/Infrastructure/API/Controller'
         tags: ['controller.service_arguments']
 
     # add more service definitions when explicit configuration is needed
