+++ back/src/Infrastructure/API/Controller/User/ListController.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace App\Infrastructure\Controller\User;
+namespace App\Infrastructure\API\Controller\User;
 
 use App\Application\Query\GetUserList;
 use App\Application\Query\GetUserListHandler;
